+++ crates/biome_service/src/configuration/json.rs
@@ -1,6 +1,7 @@
 use crate::configuration::{deserialize_line_width, serialize_line_width, PlainIndentStyle};
 use biome_deserialize_macros::{Deserializable, Merge, Partial};
 use biome_formatter::{LineEnding, LineWidth};
+use biome_json_formatter::context::TrailingCommas;
 use bpaf::Bpaf;
 use serde::{Deserialize, Serialize};
 
@@ -67,6 +68,14 @@ pub struct JsonFormatter {
     ))]
     #[partial(bpaf(long("json-formatter-line-width"), argument("NUMBER"), optional))]
     pub line_width: Option<LineWidth>,
+
+    /// Print trailing commas wherever possible in multi-line comma-separated syntactic structures. Defaults to "omit".
+    #[partial(bpaf(
+        long("json-formatter-trailing-commas"),
+        argument("omit|allow"),
+        optional
+    ))]
+    pub trailing_commas: Option<TrailingCommas>,
 }
 
 impl PartialJsonFormatter {
@@ -78,6 +87,7 @@ impl PartialJsonFormatter {
             indent_size: self.indent_size,
             line_ending: self.line_ending,
             line_width: self.line_width,
+            trailing_commas: self.trailing_commas,
         }
     }
 }
@@ -91,6 +101,7 @@ impl Default for JsonFormatter {
             indent_size: Default::default(),
             line_ending: Default::default(),
             line_width: Default::default(),
+            trailing_commas: Default::default(),
         }
     }
 }