+++ crates/biome_service/src/workspace/server.rs
@@ -377,7 +377,11 @@ impl Workspace for WorkspaceServer {
     /// Add a new file to the workspace
     fn open_file(&self, params: OpenFileParams) -> Result<(), WorkspaceError> {
         let mut file_sources = self.file_sources.write().unwrap();
-        let (index, _) = file_sources.insert_full(params.document_file_source);
+        let (index, _) = file_sources.insert_full(
+            params
+                .document_file_source
+                .unwrap_or(DocumentFileSource::from_path(&params.path)),
+        );
         self.syntax.remove(&params.path);
         self.documents.insert(
             params.path,