+++ packages/@biomejs/backend-jsonrpc/src/workspace.ts
@@ -389,6 +389,10 @@ export interface PartialJsonFormatter {
 	 * What's the max width of a line applied to JSON (and its super languages) files. Defaults to 80.
 	 */
 	lineWidth?: LineWidth;
+	/**
+	 * Print trailing commas wherever possible in multi-line comma-separated syntactic structures. Defaults to "omit".
+	 */
+	trailingCommas?: TrailingCommas;
 }
 /**
  * Options that changes how the JSON parser behaves
@@ -464,6 +468,7 @@ export type Semicolons = "always" | "asNeeded";
  * Print trailing commas wherever possible in multi-line comma-separated syntactic structures.
  */
 export type TrailingComma = "all" | "es5" | "none";
+export type TrailingCommas = "none" | "all";
 /**
  * A list of rules that belong to this group
  */