+++ CHANGELOG.md
@@ -619,6 +619,10 @@ our [guidelines for writing a good changelog entry](https://github.com/biomejs/b
 
 ### Formatter
 
+#### New features
+
+- Add option `json.formatter.trailingComma`, to provide a better control over the trailing comma in JSON/JSONC files. Its default value is `"none"`.
+
 #### Bug fixes
 
 - Fix [#1178](https://github.com/biomejs/biome/issues/1178), where the line ending option wasn't correctly applied.